1.添加AutoMapper映射帮助类 2.修改基类的一些方法名称 3.修改组织机构查询接收参数的类型 4.添加删除角色的接口 5.添加删除和修改角色的接口 6.添加根据ID获取当前节点下全部子集单位ID的方法

+++ IRepository/IOrgRepository.cs
@@ -10,7 +10,7 @@ namespace IRepository
 {
     public interface IOrgRepository:IBaseRepository<orginfo>
     {
-        List<orginfo> GetLazyOrgTree(int? ParentId,int page=1,int limit=1000,int total=0);
+        List<orginfo> GetLazyOrgTree(decimal? ParentId,int page=1,int limit=1000,int total=0);
 
         DataTable GetOrgTree();
